refactor(appointment): 날짜 및 시간을 포맷팅하는 함수를 유틸화
- 중복으로 사용해야 하는 경우가 생겨서, 날짜 및 시간을 포맷팅하는 getFormattedDateTime 함수를 유틸화
